Adds infrastructure to support developing using boot
Sometimes, in the course of development, we want to make changes to libraries
we use. In any given boot project's `build.boot` file, we can specify a
`checkouts` environment key which allows us to point to jars in our own maven
repositories. Bootlaces provides several convenience functions for facilitating
the building of these jars and installing them into our own maven repositories
for access in our other projects.

This commit introduces the infrastructure necessary to make use of bootlaces'
functions--specifically `build-jar`--and paves the way for publishing om to
clojars as an alternative to leiningen. (Of course, this is entirely up to the
maintainer(s) of this project.)